Changes made: * Split up src/NFLLunch/app into app, models, service * Add __init__ to src/NFLLunch * Update requirements
