+++ setup.py
@@ -36,6 +36,7 @@ setup(
         "click",
         "flask-rebar",
         "flask-marshmallow",
+        "sleeper-api-wrapper",
         "nflgame-redux==2.0.1b1"
     ],
 )
